reorderung: moving implementation files into separate directory tree
