+++ Doxyfile
@@ -666,7 +666,7 @@ WARN_LOGFILE           =
 # directories like "/usr/src/myproject". Separate the files or directories
 # with spaces.
 
-INPUT                  = include tests examples doc/source README.md CHANGELOG.md
+INPUT                  = include src tests examples doc/source README.md CHANGELOG.md
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding, which is
