+++ include/pfasst/controller.hpp
@@ -1,6 +1,6 @@
 #ifndef _PFASST__CONTROLLER_HPP
 #define _PFASST__CONTROLLER_HPP
 
-#include "controller/controller.hpp"
+#include "pfasst/controller/controller.hpp"
 
 #endif  // _PFASST__CONTROLLER_HPP
