+++ src/pfasst/controller/sdc_impl.hpp
@@ -1,4 +1,4 @@
-#include "sdc.hpp"
+#include "pfasst/controller/sdc.hpp"
 
 
 namespace pfasst
