+++ src/pfasst/quadrature/clenshaw_curtis_impl.hpp
@@ -1,7 +1,6 @@
-#include "uniform.hpp"
+#include "pfasst/quadrature/clenshaw_curtis.hpp"
 
 #include <stdexcept>
-#include <vector>
 using namespace std;
 
 
@@ -10,34 +9,36 @@ namespace pfasst
   namespace quadrature
   {
     template<typename precision>
-    Uniform<precision>::Uniform(const size_t num_nodes)
+    ClenshawCurtis<precision>::ClenshawCurtis(const size_t num_nodes)
       : IQuadrature<precision>(num_nodes)
     {
       if (this->num_nodes < 2) {
-        throw invalid_argument("Uniform quadrature requires at least two quadrature nodes.");
+        throw invalid_argument("Clenshaw-Curtis quadrature requires at least two quadrature nodes.");
       }
       this->compute_nodes();
       this->compute_weights();
     }
 
     template<typename precision>
-    bool Uniform<precision>::left_is_node() const
+    bool ClenshawCurtis<precision>::left_is_node() const
     {
       return LEFT_IS_NODE;
     }
 
     template<typename precision>
-    bool Uniform<precision>::right_is_node() const
+    bool ClenshawCurtis<precision>::right_is_node() const
     {
       return RIGHT_IS_NODE;
     }
 
     template<typename precision>
-    void Uniform<precision>::compute_nodes()
+    void ClenshawCurtis<precision>::compute_nodes()
     {
       this->nodes = vector<precision>(this->num_nodes, precision(0.0));
+      auto roots = Polynomial<precision>::legendre(this->num_nodes).roots();
+
       for (size_t j = 0; j < this->num_nodes; j++) {
-        this->nodes[j] = precision(j) / (this->num_nodes - 1);
+        this->nodes[j] = 0.5 * (1.0 - cos(j * pi<precision>() / (this->num_nodes - 1)));
       }
     }
   }  // ::pfasst::quadrature
