+++ src/pfasst/quadrature/gauss_legendre_impl.hpp
@@ -1,7 +1,9 @@
-#include "clenshaw_curtis.hpp"
+#include "pfasst/quadrature/gauss_legendre.hpp"
 
-#include <stdexcept>
+#include <vector>
 using namespace std;
+
+#include "pfasst/quadrature/polynomial.hpp"
 
 
 namespace pfasst
@@ -9,36 +11,33 @@ namespace pfasst
   namespace quadrature
   {
     template<typename precision>
-    ClenshawCurtis<precision>::ClenshawCurtis(const size_t num_nodes)
+    GaussLegendre<precision>::GaussLegendre(const size_t num_nodes)
       : IQuadrature<precision>(num_nodes)
     {
-      if (this->num_nodes < 2) {
-        throw invalid_argument("Clenshaw-Curtis quadrature requires at least two quadrature nodes.");
-      }
       this->compute_nodes();
       this->compute_weights();
     }
 
     template<typename precision>
-    bool ClenshawCurtis<precision>::left_is_node() const
+    bool GaussLegendre<precision>::left_is_node() const
     {
       return LEFT_IS_NODE;
     }
 
     template<typename precision>
-    bool ClenshawCurtis<precision>::right_is_node() const
+    bool GaussLegendre<precision>::right_is_node() const
     {
       return RIGHT_IS_NODE;
     }
 
     template<typename precision>
-    void ClenshawCurtis<precision>::compute_nodes()
+    void GaussLegendre<precision>::compute_nodes()
     {
       this->nodes = vector<precision>(this->num_nodes, precision(0.0));
       auto roots = Polynomial<precision>::legendre(this->num_nodes).roots();
 
       for (size_t j = 0; j < this->num_nodes; j++) {
-        this->nodes[j] = 0.5 * (1.0 - cos(j * pi<precision>() / (this->num_nodes - 1)));
+        this->nodes[j] = 0.5 * (1.0 + roots[j]);
       }
     }
   }  // ::pfasst::quadrature
